+++ js/tests/test_timepicker.js
@@ -1,7 +1,6 @@
 import {TimepickerWidget} from '../src/timepicker.js';
 import {register_timepicker_array_subscribers} from '../src/timepicker.js';
 
-window.yafowil_array = undefined;
 
 QUnit.module('TimepickerWidget', hooks => {
     let container = $('<div id="container" />');